Split out swing example from swing library
Fixes #13.  cleans up a few more bits of sampleGrids and voxels as well.
