+++ swing-demo/build.gradle
@@ -1,0 +1,8 @@
+apply plugin: 'application'
+
+mainClassName = 'wannabe.swing.SwingWannabe'
+
+dependencies {
+	compile project(':library')
+	compile project(':swing')
+}
