+++ swing-demo/pom.xml
@@ -1,0 +1,25 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>muddyhorse.Wannabe</groupId>
+    <artifactId>parent</artifactId>
+    <version>0.0.1-SNAPSHOT</version>
+  </parent>
+  <artifactId>swing-demo</artifactId>
+  <packaging>jar</packaging>
+  <name>Wannabe-Swing Demo</name>
+  <description>Wannabe demo in Swing</description>
+  
+  <dependencies>
+  <dependency>
+    <groupId>muddyhorse.Wannabe</groupId>
+    <artifactId>library</artifactId>
+    <version>0.0.1-SNAPSHOT</version>
+  </dependency>
+  <dependency>
+    <groupId>muddyhorse.Wannabe</groupId>
+    <artifactId>swing</artifactId>
+    <version>0.0.1-SNAPSHOT</version>
+  </dependency>
+  </dependencies>
+</project>
