+++ swing-demo/src/main/resources/example-heightmap.png
